Test: 강의의 가격과 결제 금액이 같을 때, 수강신청이 된다

+++ src/main/java/nextstep/sessions/domain/SessionDetails.java
@@ -35,7 +35,7 @@ public class SessionDetails {
         this.currentCountOfStudents++;
     }
 
-    public boolean isSamePrice(int amount) {
+    public boolean isSamePrice(long amount) {
         return this.price == amount;
     }
